Add a new song jump menu
