+++ src/JumpMenu/Main.tsx
@@ -4,6 +4,7 @@ import { Button, Flex } from 'rebass'
 import { useJumpNavigationContext } from './JumpNavigationContextProvider'
 import { useInputContext } from './InputContextProvider'
 
+import AddANewSong from './Menus/AddANewSong'
 import AllTags from './Menus/AllTags'
 import FindInLibrary from './Menus/FindInLibrary'
 import Preview from './Menus/Preview'
@@ -18,6 +19,7 @@ const Main: React.FC = () => {
 
   const Route = {
     allTags: AllTags,
+    externalSearch: AddANewSong,
     home: Home,
     keyboardShortcuts: KeyboardShortcuts,
     musicboxSearch: FindInLibrary,
@@ -32,7 +34,7 @@ const Main: React.FC = () => {
   return (
     <>
       <Route />
-      <Flex alignItems="center" justifyContent="space-between" flexDirection="row-reverse">
+      <Flex alignItems="center" justifyContent="space-between" flexDirection="row-reverse" my={2}>
         <Button onClick={close}>Close</Button>
         {selectedMenu !== 'home' && <Button onClick={back}>Back</Button>}
       </Flex>
