+++ src/JumpMenu/Menus/AddANewSong/index.ts
@@ -1,0 +1,3 @@
+import AddANewSong from './AddANewSong'
+
+export default AddANewSong
