--- src/JumpMenu/Results/index.ts
@@ -1,4 +1,0 @@
-import Results from './Results'
-
-export * from './deserialize'
-export default Results
